Added trip skeleton Changed trains
