+++ src/models/trainModel.go
@@ -9,8 +9,8 @@ import (
 )
 
 const (
-	saveToTrip  = "INSERT INTO trips_trains (trips_id, trains_id) VALUES ($1, $2)"
-	getFromTrip = "SELECT * FROM trains INNER JOIN trips_trains ON trips_trains.trains_id = trains.id AND trips_trains.trips_id = $1"
+	saveTrainToTrip  = "INSERT INTO trips_trains (trip_id, train_id) VALUES ($1, $2)"
+	getTrainFromTrip = "SELECT * FROM trains INNER JOIN trips_trains ON trips_trains.train_id = trains.id AND trips_trains.trip_id = $1"
 )
 
 //Task representation in DB
@@ -75,14 +75,14 @@ var GetTrains = func(params url.Values) ([]Train, error) {
 
 //SaveTrain used for saving trains to Trip
 var SaveTrain = func(tripsID, trainsID uuid.UUID) error {
-	_, err := database.DB.Exec(saveToTrip, tripsID, trainsID)
+	_, err := database.DB.Exec(saveTrainToTrip, tripsID, trainsID)
 
 	return err
 }
 
-//GetFromTrip used for getting trains from Trip
-var GetFromTrip = func(tripsID uuid.UUID) ([]Train, error) {
-	rows, err := database.DB.Query(getFromTrip, tripsID)
+//GetTrainFromTrip used for getting trains from Trip
+var GetTrainFromTrip = func(tripsID uuid.UUID) ([]Train, error) {
+	rows, err := database.DB.Query(getTrainFromTrip, tripsID)
 	if err != nil {
 		return []Train{}, err
 	}
